FIX: fixed layout wrapper prop errors and removed provider
