+++ app/(main)/(guest)/guest/layout.tsx
@@ -1,8 +1,21 @@
 import "@/app/globals.css";
-import GuestLayout from "@/components/Provider/GuestLayout";
+import SideNavBar from "@/components/Shared/SideNavBar";
 
-function GuestLayoutWrapper({ children }: { children: React.ReactNode }) {
-  return <GuestLayout children={children} />;
+export default async function GuestLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const isGuest = true;
+
+  return (
+    <div className="flex h-screen w-full overflow-x-hidden">
+      <SideNavBar isGuest={isGuest} />
+      <div className="flex flex-col w-full relative h-screen bg-brand-ash  overflow-y-scroll">
+        <div className="lg:mt-14 w-full p-4 h-[calc(100vh-80px)] ">
+          {children}
+        </div>
+      </div>
+    </div>
+  );
 }
-
-export default GuestLayoutWrapper;
